requirements.txt
asgiref==3.3.4
backcall==0.2.0
certifi==2020.12.5
decorator==4.4.2
Django==3.2
fpdf==1.7.2
image==1.5.33
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
jedi==0.17.0
jupyter-client==6.1.3
jupyter-core==4.6.3
opencv-contrib-python==4.4.0.42
parso==0.7.0
pickleshare==0.7.5
Pillow==8.2.0
prompt-toolkit==3.0.5
Pygments==2.6.1
python-dateutil==2.8.1
pytz==2021.1
pyzmq==19.0.1
six==1.15.0
sqlparse==0.4.1
tornado==6.0.4
traitlets==4.3.3
wcwidth==0.2.4
wincertstore==0.2
